chore: make scale down lambda common
