+++ perf/terraform/.gitignore
@@ -34,3 +34,4 @@ override.tf.json
 terraform.rc
 
 *.pub
+scale_down.zip
