+++ perf/terraform/common/files/scale_down.py
@@ -1,0 +1,28 @@
+import boto3
+import os
+import json
+import datetime
+
+regions = os.environ['AWS_REGIONS']
+tags = os.environ['TAGS']
+max_age_minutes = os.environ['MAX_AGE_MINUTES']
+
+
+def is_lost(instance):
+    return instance.launch_time.replace(tzinfo=None) < datetime.datetime.now(datetime.timezone.utc) - datetime.timedelta(minutes=max_age_minutes)
+
+
+def lambda_handler(event, context):
+    # iterate over all regions
+    for region in regions.split(','):
+        ec2 = boto3.client('ec2', region_name=region)
+        filter = [{'Name': 'instance-state-name', 'Values': ['running']}]
+        filter = filter + [{
+            'Name': 'tag:' + k,
+            'Values': [v]
+        } for k, v in json.loads(tags).items()]
+        instances = [i.id for i in ec2.describe_instances(
+            Filters=filter) if is_lost(i)]
+        print('found instances: ' + str(instances))
+        ec2.stop_instances(InstanceIds=instances)
+        print('stopped instances: ' + str(instances))
